refactor: NewsFragment, NoticeUiModel 경로 변경 및 관련 코드 수정
- NewsFragment: presentation.news.notice -> presentation.news
- NoticeUiModel: presentation.news.notice -> presentation.news.notice.model
- PlaceListViewHolder, PlaceListAdapter: presentation.placeList -> presentation.placeList.adapter
- 레이아웃 파일의 tools:context 및 data binding 경로 수정
- MainActivity의 NewsFragment import 경로 수정
- 레이아웃 파일 내 뷰 속성 순서 변경
