+++ android/app/src/main/java/com/daedan/festabook/presentation/news/NewsFragment.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.news.notice
+package com.daedan.festabook.presentation.news
 
 import android.os.Bundle
 import android.view.View
@@ -6,6 +6,7 @@ import com.daedan.festabook.R
 import com.daedan.festabook.databinding.FragmentNewsBinding
 import com.daedan.festabook.presentation.common.BaseFragment
 import com.daedan.festabook.presentation.news.notice.adapter.NoticeAdapter
+import com.daedan.festabook.presentation.news.notice.model.NoticeUiModel
 
 class NewsFragment : BaseFragment<FragmentNewsBinding>(R.layout.fragment_news) {
     private val noticeAdapter: NoticeAdapter by lazy { NoticeAdapter() }
