+++ android/app/src/main/java/com/daedan/festabook/presentation/news/notice/adapter/NoticeAdapter.kt
@@ -3,7 +3,7 @@ package com.daedan.festabook.presentation.news.notice.adapter
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import com.daedan.festabook.presentation.news.notice.NoticeUiModel
+import com.daedan.festabook.presentation.news.notice.model.NoticeUiModel
 
 class NoticeAdapter : ListAdapter<NoticeUiModel, NoticeViewHolder>(noticeDiffCallback) {
     override fun onCreateViewHolder(
