+++ android/app/src/main/java/com/daedan/festabook/presentation/placeList/PlaceListFragment.kt
@@ -5,6 +5,7 @@ import android.view.View
 import com.daedan.festabook.R
 import com.daedan.festabook.databinding.FragmentPlaceListBinding
 import com.daedan.festabook.presentation.common.BaseFragment
+import com.daedan.festabook.presentation.placeList.adapter.PlaceListAdapter
 import com.daedan.festabook.presentation.placeList.dummy.DummyPlace
 
 class PlaceListFragment :
